# openapi.yaml
openapi: 3.0.3
info:
  title: Finanzas Backend API
  description: Personal finance ETL + API
  version: 0.3.0
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /files:
    get:
      tags: [files]
      summary: List uploaded files for the current user
      operationId: listFiles
      security:
        - BearerAuth: []
      parameters:
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order by uploaded_at (default newest first)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileMetadata'
  /files/{file_id}/preview:
    get:
      tags: [files]
      summary: Preview raw rows of a processed file
      operationId: previewFile
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FileId'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of rows to return
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilePreviewResponse'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /files/{file_id}/url:
    get:
      tags: [files]
      summary: Get a signed URL for direct file access
      description: |
        Returns a short-lived Supabase Storage signed URL so the client can
        open or download the file directly without proxying through the backend.
        Only available when Supabase Storage is configured (returns 501 otherwise).
      operationId: getFileUrl
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FileId'
        - name: expires_in
          in: query
          required: false
          schema:
            type: integer
            minimum: 60
            maximum: 86400
            default: 3600
          description: URL validity in seconds (default 1 hour, max 24 hours)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  file_id:
                    type: string
                    format: uuid
                  url:
                    type: string
                    format: uri
                    description: Signed URL valid for `expires_in` seconds
                  expires_in:
                    type: integer
                    description: Validity duration in seconds
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '501':
          description: Supabase Storage not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /files/{file_id}:
    delete:
      tags: [files]
      summary: Delete a source file
      description: |
        Deletes the file from Supabase Storage and removes the database record.
        Returns 409 if the file has associated transactions — run
        `DELETE /etl/reset/{file_id}` first to remove them.
      operationId: deleteFile
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '204':
          description: Deleted
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: File has associated transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /files/upload:
    post:
      tags: [files]
      summary: Upload a bank statement file
      description: |
        Accepts `.xlsx` or `.pdf` files for any supported bank.
        Returns the existing record if the file hash already exists.
        If `file_password` is provided, the PDF is unlocked in memory before
        being stored — the password itself is never persisted.
        The `account_id` must belong to the authenticated user.
      operationId: uploadFile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, account_id]
              properties:
                file:
                  type: string
                  format: binary
                  description: Bank statement file (.xlsx or .pdf)
                account_id:
                  type: string
                  format: uuid
                  description: Account this file belongs to
                file_password:
                  type: string
                  description: |
                    Optional password for password-protected files (e.g. Nequi PDF uses the account holder's cédula).
                    Used to unlock the PDF before storage — never persisted.
      responses:
        '200':
          description: OK (file already existed or newly created)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
        '400':
          description: Unsupported file type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'

  # ──────────────────────────────────────────────
  # ETL
  # ──────────────────────────────────────────────
...
  /transactions:
    get:
      tags: [transactions]
      summary: List transactions for the current user
      operationId: listTransactions
      security:
        - BearerAuth: []
      parameters:
        - name: account_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter by account UUID
        - $ref: '#/components/parameters/OwnerFilter'
        - $ref: '#/components/parameters/DateFromFilter'
        - $ref: '#/components/parameters/DateToFilter'
        - $ref: '#/components/parameters/CategoryFilter'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'
  /transactions/{transaction_id}/categorize:
    patch:
      tags: [transactions]
      summary: Manually recategorize a transaction
      description: |
        Sets the category on the transaction with `category_method=USER`.
        Optionally seeds a new `category_example` from `description_clean`
        to improve future automatic categorization.
      operationId: recategorizeTransaction
      security:
        - BearerAuth: []
      parameters:
        - name: transaction_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecategorizeRequest'
            example:
              category: OCIO
              description_clean: NETFLIX DL
      responses:
        '200':
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'

  # ──────────────────────────────────────────────
  # DASHBOARD
  # ──────────────────────────────────────────────
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    FileId:
      name: file_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Source file UUID
...
    OwnerFilter:
      name: owner
      in: query
      required: false
      schema:
        type: string
      description: Filter by owner label (e.g. "PACHO", "LU")
...
    DateFromFilter:
      name: date_from
      in: query
      required: false
      schema:
        type: string
        format: date
        example: "2026-01-01"
      description: Start date (inclusive) in `YYYY-MM-DD` format
    DateToFilter:
      name: date_to
      in: query
      required: false
      schema:
        type: string
        format: date
        example: "2026-03-31"
      description: End date (inclusive) in `YYYY-MM-DD` format
    CategoryFilter:
      name: category
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/Category'
      description: Filter by transaction category
  schemas:
...
    BankEnum:
      type: string
      enum: [BANCOLOMBIA, RAPPI, FALABELLA, NEQUI]
    AccountTypeEnum:
      type: string
      enum: [CREDITO, DEBITO, AHORROS]
    Category:
      type: string
      enum:
        - HOGAR
        - DOMICILIOS
        - CARRO
        - TRANSPORTE
        - OCIO
        - RESTAURANTES
        - ROPA
        - SALUD
        - PRESTACIONES
        - REGALOS
        - EDUCACION
        - TRABAJO
        - COBRO_BANCARIO
        - PAGO
        - PLATAFORMAS
        - INGRESO
        - INVERSION
        - MOVIMIENTO_ENTRE_BANCOS
        - OTROS
    CategoryMethod:
      type: string
      enum: [RULES, LLM, USER]
    ParseStatus:
      type: string
      enum: [UPLOADED, PROCESSED, FAILED]

    # ── Auth ───────────────────────────────────
...
    FileMetadata:
      type: object
      properties:
        file_id:
          type: string
          format: uuid
        filename:
          type: string
        account_id:
          type: string
          format: uuid
        account_name:
          type: string
        status:
          type: string
          enum: [UPLOADED, PARSING, SUCCESS, FAILED]
          description: Current processing status of the file
        uploaded_at:
          type: string
          format: date-time
        hash:
          type: string
          description: SHA-256 hash of the file
    FilePreviewResponse:
      type: object
      properties:
        file_id:
          type: string
          format: uuid
        filename:
          type: string
        columns:
          type: array
          items:
            type: string
          description: Column names derived from the raw parsed data
        rows:
          type: array
          items:
            type: object
          description: First N raw rows as parsed from the source file
        total_rows:
          type: integer
          description: Total number of raw rows stored for this file
    FileUploadResponse:
      type: object
      properties:
        file_id:
          type: string
          format: uuid
        parse_status:
          $ref: '#/components/schemas/ParseStatus'
        file_hash:
          type: string
          description: SHA-256 hash of the file (used for deduplication)
        original_filename:
          type: string
        uploaded_at:
          type: string
          format: date-time

    # ── ETL ────────────────────────────────────
...
    TransactionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        source_file_id:
          type: string
          format: uuid
        posted_at:
          type: string
          format: date
        description_raw:
          type: string
        description_clean:
          type: string
        amount:
          type: number
          description: Amount in COP (USD transactions are converted at 4000 COP/USD). Negative = expense, positive = income.
        currency:
          type: string
          example: COP
        merchant_guess:
          type: string
          nullable: true
        details_json:
          type: object
          nullable: true
        category:
          $ref: '#/components/schemas/Category'
          nullable: true
        category_confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          nullable: true
        category_method:
          $ref: '#/components/schemas/CategoryMethod'
          nullable: true
        created_at:
          type: string
          format: date-time
        account_id:
          type: string
          format: uuid
        account_name:
          type: string
        bank_name:
          $ref: '#/components/schemas/BankEnum'
        owner:
          type: string
          description: Free-form owner label on the account
        account_type:
          $ref: '#/components/schemas/AccountTypeEnum'
    RecategorizeRequest:
      type: object
      required: [category]
      properties:
        category:
          $ref: '#/components/schemas/Category'
        description_clean:
          type: string
          nullable: true
          description: |
            If provided, seeds a new category_example entry so future
            transactions with this description are auto-categorized.

    # ── Dashboard ──────────────────────────────
...
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string